+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 3.15.2 (August 8, 2022)
+
+### Added
+
+- Add `boundary-x` and `boundary-y` options to Drop component
+
+## Fixed
+
+- Remove click event listener in Scroll component correctly 
+
 ## 3.15.1 (July 15, 2022)
 
 ### Changed